refactor(electron): split external player sessions
